+++ uasset/src/uasset/unreal_types.rs
@@ -3,10 +3,10 @@ pub type Guid = [u8; 16];
 
 pub fn new_guid(a: u32, b: u32, c: u32, d: u32) -> Guid {
     [
-        a & 0xff, (a >> 8) & 0xff, (a >> 16) & 0xff, (a >> 24) & 0xff,
-        b & 0xff, (b >> 8) & 0xff, (b >> 16) & 0xff, (b >> 24) & 0xff,
-        c & 0xff, (c >> 8) & 0xff, (c >> 16) & 0xff, (c >> 24) & 0xff,
-        d & 0xff, (d >> 8) & 0xff, (d >> 16) & 0xff, (d >> 24) & 0xff
+        (a & 0xff) as u8, ((a >> 8) & 0xff) as u8, ((a >> 16) & 0xff) as u8, ((a >> 24) & 0xff) as u8,
+        (b & 0xff) as u8, ((b >> 8) & 0xff) as u8, ((b >> 16) & 0xff) as u8, ((b >> 24) & 0xff) as u8,
+        (c & 0xff) as u8, ((c >> 8) & 0xff) as u8, ((c >> 16) & 0xff) as u8, ((c >> 24) & 0xff) as u8,
+        (d & 0xff) as u8, ((d >> 8) & 0xff) as u8, ((d >> 16) & 0xff) as u8, ((d >> 24) & 0xff) as u8
     ]
 }
 
@@ -16,7 +16,7 @@ pub struct GenerationInfo {
     name_count: i32,
 }
 
-#[derive(Debug, Default)]
+#[derive(Debug, Default, Hash, PartialEq, Eq)]
 pub struct FName {
     pub content: String,
     pub index: i32
@@ -30,7 +30,7 @@ impl FName {
     }
 }
 
-#[derive(Debug)]
+#[derive(Debug, Default)]
 pub struct NamespacedString {
     pub namespace: String,
     pub value: String
